refactor(ui): extract declarative handler system and service worker from app.js
Move the event delegation infrastructure (invokeBoundExpression,
bindDeclarativeHandlers, debounce, DEBOUNCED_INPUT_EXPRESSIONS) and
service worker registration into client/bootstrap/initGlobalListeners.js.

app.js now imports and calls these as part of its initialization sequence.
This is the first step toward making app.js a thin composition root —
domain-specific helpers will move to feature initializers in subsequent PRs.

app.js: 1,797 → 1,699 lines (~100 lines extracted)

Closes #386
